requirements.txt
flask>=2.3.0
openai
python-dotenv>=1.0.0
flask-limiter>=3.5.0
markdown>=3.5.0